nutr_info to torch.float32

+++ food_dataset.py
@@ -32,7 +32,7 @@ class FoodDataset(Dataset):
     def __getitem__(self, idx):
         sample = dict(self.nutr_info[['kcal_100', 'mass', 'prot_100', 
                                       'fat_100', 'carb_100']].iloc[idx])
-        sample['nutr_info'] = torch.tensor(list(sample.values()))
+        sample['nutr_info'] = torch.tensor(list(sample.values())).to(torch.float32)
         if self.is_read_all:
             image = self.images[idx]
         else:
